requirements.txt
altair==5.5.0
attrs==25.3.0
blinker==1.9.0
brotlicffi @ file:///croot/brotlicffi_1736182461069/work
cachetools==6.2.0
certifi @ file:///croot/certifi_1754570635119/work/certifi
cffi @ file:///croot/cffi_1736182485317/work
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click==8.3.0
contourpy==1.3.3
cycler==0.12.1
filelock @ file:///croot/filelock_1744281381737/work
fonttools==4.60.1
fsspec==2024.6.1
gitdb==4.0.12
GitPython==3.1.45
gmpy2 @ file:///croot/gmpy2_1738085463648/work
hf-xet==1.1.10
huggingface-hub==0.35.3
idna @ file:///croot/idna_1714398848350/work
Jinja2 @ file:///croot/jinja2_1741710844255/work
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
MarkupSafe @ file:///croot/markupsafe_1738584038848/work
matplotlib==3.10.6
mkl-service==2.4.0
mkl_fft @ file:///io/mkl313/mkl_fft_1730824109137/work
mkl_random @ file:///io/mkl313/mkl_random_1730823916628/work
mpmath @ file:///croot/mpmath_1690848262763/work
narwhals==2.6.0
networkx @ file:///croot/networkx_1756709277056/work
numpy @ file:///croot/numpy_and_numpy_base_1725470312869/work/dist/numpy-2.0.1-cp311-cp311-linux_x86_64.whl#sha256=372b9067cf9c96d8662c6f20ef49eeda43f0f8d2f074820224c954d8d8fae7bc
packaging==25.0
pandas==2.3.3
pillow @ file:///croot/pillow_1752524526621/work
protobuf==6.32.1
psutil==7.1.0
pyarrow==21.0.0
pycparser @ file:///home/task_175749606196006/conda-bld/pycparser_1757496076770/work
pydeck==0.9.1
pyparsing==3.2.5
PySocks @ file:///work/ci_py311/pysocks_1676822712504/work
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML @ file:///croot/pyyaml_1728657952215/work
referencing==0.36.2
regex==2025.9.18
requests @ file:///croot/requests_1756709366904/work
rpds-py==0.27.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
six==1.17.0
smmap==5.0.2
streamlit==1.50.0
sympy==1.13.1
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.22.1
toml==0.10.2
torch==2.5.1
torchaudio==2.5.1
torchsummary==1.5.1
torchvision==0.20.1
tornado==6.5.2
tqdm==4.67.1
triton==3.1.0
typing_extensions @ file:///croot/typing_extensions_1756280817316/work
tzdata==2025.2
urllib3 @ file:///croot/urllib3_1750775463400/work
watchdog==6.0.0
